requirements.txt
amqp==1.4.8
anyjson==0.3.3
beautifulsoup4==4.4.1
billiard==3.3.0.22
bottlenose==0.6.2
celery==3.1.19
decorator==4.0.6
defusedxml==0.4.1
Django==1.8.7
django-appconf==1.0.1
django-celery==3.1.17
django-debug-toolbar==1.4
django-email-html==0.1.9
django-ipware==1.1.2
django-js-reverse==0.6.1
django-model-utils==2.3.1
django-notifications-hq==0.8.0
django-user-accounts==1.3.1
django-widget-tweaks==1.4.1
fake-factory==0.5.3
geoip2==2.2.0
goodreads==0.2.4
httplib2==0.9.2
ipython-genutils==0.1.0
jsonfield==1.0.3
kombu==3.0.32
maxminddb==1.2.0
mixer==5.3.1
nose==1.3.7
numpy==1.10.1
oauth2==1.9.0.post1
oauthlib==1.0.3
pandas==0.17.1
path.py==8.1.2
pexpect==4.0.1
pickleshare==0.5
psycopg2==2.6.1
ptyprocess==0.5
PyJWT==1.4.0
python-dateutil==2.4.2
python-openid==2.2.5
python-social-auth==0.2.13
python3-openid==3.0.9
pytz==2015.7
rauth==0.7.2
requests==2.9.0
requests-oauthlib==0.6.0
retrying==1.3.3
simplegeneric==0.8.1
six==1.10.0
sqlparse==0.1.18
traitlets==4.0.0
wheel==0.26.0
xmltodict==0.9.2
